requirements.txt
numpy==1.16.1
torch==1.0.1
pillow==5.3.0
torchvision==0.2.1
tqdm==4.31.0
